requirements.txt
python-digitalocean==1.17.0
python-dotenv==1.0.0
mcstatus==11.1.1
